backend/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.41
psycopg==3.1.18
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.30.0
httpx==0.27.0
tenacity==8.2.3
fastapi-crudrouter==0.8.6
sqlmodel==0.0.16
python-multipart==0.0.6
openai==1.35.1
google-auth==2.31.0
requests==2.32.3
jinja2==3.1.4
json-fix==1.0.0
types-jinja2==2.11.9
pytest==8.2.2
respx==0.21.1
litellm==1.30.0
alembic
psycopg2-binary
